Composer/actions: simplify the PHPCS threshold check
As of YoastCS 2.2.0, the Yoast standard contains a custom `Threshold` report.

This switches out the custom logic for comparing the PHPCS run results against thresholds in the `Yoast\WP\SEO\Composer\Actions` class to use this new report instead.

The report will now also respect the error code returned by PHPCS if the threshold is not met.

Notes:
* The `-m` parameter directs PHPCS to not record the error messages, which saves a lot of memory usage during this run.
* The -q parameter makes it so no progress is being shown during this run.
* The `exec()` function has been switched out in favour of the `passthru()` to allow for the report to be displayed directly.
    The downside of this is that the progress of the run will also be displayed, but IMO that's a small price to pay for the convenience of not having to go through the output and manually display the report.