WAIT tagged release | PHP 8.0 | Update OAuth PHP Scoper config
PHP Scoper is prefixing the attribute use statement that will need to be pulled into the TheLeague OAuth Client implenetation of the JSONSerializableInterface.
See https://github.com/thephpleague/oauth2-client/pull/919

This change removes the prefix from the use statement.

PHP Scoper was updated to 0.15.0 during investigation of this issue and was not worked as desired on the newer version. Bumping the version in this commit seemed out of scope.

This is addressed in the following ticket: https://github.com/humbug/php-scoper/issues/539